makes express app exportable for testing
